+++ source/Settings.cpp
@@ -592,6 +592,7 @@ namespace settings
 			{
 				using namespace controls;
 				hideKeyCode = Read<std::int32_t>(c, "iHideKeyCode:Controls", hideKeyCode);
+				holdToPanSecs = Read<float>(c, "fHoldToPanSecs:Controls", holdToPanSecs);
 				zoomToggleKeyCode = Read<std::int32_t>(c, "iZoomToggleKeyCode:Controls", zoomToggleKeyCode);
 				zoomDefault = Read<float>(c, "fZoomDefault:Controls", zoomDefault);
 				zoomZoomedIn = Read<float>(c, "fZoomZoomedIn:Controls", zoomZoomedIn);
@@ -642,6 +643,7 @@ namespace settings
 		{
 			using namespace controls;
 			add("iHideKeyCode:Controls", static_cast<int>(hideKeyCode));
+			add("fHoldToPanSecs:Controls", holdToPanSecs);
 			add("iZoomToggleKeyCode:Controls", static_cast<int>(zoomToggleKeyCode));
 			add("fZoomDefault:Controls", zoomDefault);
 			add("fZoomZoomedIn:Controls", zoomZoomedIn);
@@ -736,6 +738,7 @@ namespace settings
 		ok &= WriteBool(kDisplaySection, "bShowOnGameStart", display::showOnGameStart);
 
 		ok &= WriteInt(kControlsSection, "iHideKeyCode", controls::hideKeyCode);
+		ok &= WriteFloat(kControlsSection, "fHoldToPanSecs", controls::holdToPanSecs);
 		ok &= WriteInt(kControlsSection, "iZoomToggleKeyCode", controls::zoomToggleKeyCode);
 		ok &= WriteFloat(kControlsSection, "fZoomDefault", controls::zoomDefault);
 		ok &= WriteFloat(kControlsSection, "fZoomZoomedIn", controls::zoomZoomedIn);